requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.5
    # via
    #   discord-py
    #   litellm
aiosignal==1.4.0
    # via aiohttp
alembic==1.18.4
    # via mindforge (pyproject.toml)
annotated-doc==0.0.4
    # via
    #   fastapi
    #   typer
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
argon2-cffi==25.1.0
    # via minio
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
asyncpg==0.31.0
    # via mindforge (pyproject.toml)
attrs==26.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
audioop-lts==0.2.2
    # via discord-py
backoff==2.2.1
    # via langfuse
bcrypt==5.0.0
    # via mindforge (pyproject.toml)
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   minio
    #   requests
cffi==2.0.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
charset-normalizer==3.4.7
    # via requests
click==8.1.8
    # via
    #   litellm
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   uvicorn
cryptography==46.0.7
    # via pyjwt
discord-py==2.7.1
    # via mindforge (pyproject.toml)
distro==1.9.0
    # via openai
fastapi==0.135.3
    # via mindforge (pyproject.toml)
fastuuid==0.14.0
    # via litellm
filelock==3.25.2
    # via huggingface-hub
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2026.3.0
    # via huggingface-hub
googleapis-common-protos==1.74.0
    # via opentelemetry-exporter-otlp-proto-http
greenlet==3.4.0
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.4.3
    # via huggingface-hub
hiredis==3.3.1
    # via redis
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   huggingface-hub
    #   langfuse
    #   litellm
    #   mindforge (pyproject.toml)
    #   openai
huggingface-hub==1.10.1
    # via tokenizers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.5.0
    # via
    #   litellm
    #   opentelemetry-api
jinja2==3.1.6
    # via litellm
jiter==0.14.0
    # via openai
jsonschema==4.23.0
    # via litellm
jsonschema-specifications==2025.9.1
    # via jsonschema
langfuse==4.2.0
    # via mindforge (pyproject.toml)
litellm==1.83.4
    # via mindforge (pyproject.toml)
lxml==6.0.3
    # via python-docx
mako==1.3.10
    # via alembic
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
minio==7.2.20
    # via mindforge (pyproject.toml)
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
neo4j==6.1.0
    # via mindforge (pyproject.toml)
openai==2.30.0
    # via litellm
opentelemetry-api==1.41.0
    # via
    #   langfuse
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-common==1.41.0
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-http==1.41.0
    # via langfuse
opentelemetry-proto==1.41.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.41.0
    # via
    #   langfuse
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.62b0
    # via opentelemetry-sdk
packaging==26.0
    # via
    #   huggingface-hub
    #   langfuse
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.33.6
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
pycparser==3.0
    # via cffi
pycryptodome==3.23.0
    # via minio
pydantic==2.12.5
    # via
    #   fastapi
    #   langfuse
    #   litellm
    #   mindforge (pyproject.toml)
    #   openai
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via mindforge (pyproject.toml)
pygments==2.20.0
    # via rich
pyjwt[crypto]==2.12.1
    # via mindforge (pyproject.toml)
pymupdf==1.27.2.2
    # via mindforge (pyproject.toml)
python-docx==1.2.0
    # via mindforge (pyproject.toml)
python-dotenv==1.0.1
    # via
    #   litellm
    #   pydantic-settings
    #   uvicorn
python-frontmatter==1.1.0
    # via mindforge (pyproject.toml)
python-multipart==0.0.26
    # via mindforge (pyproject.toml)
pytz==2026.1.post1
    # via neo4j
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   python-frontmatter
    #   uvicorn
redis[hiredis]==7.4.0
    # via mindforge (pyproject.toml)
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2026.4.4
    # via tiktoken
requests==2.33.1
    # via
    #   opentelemetry-exporter-otlp-proto-http
    #   tiktoken
rich==15.0.0
    # via typer
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via typer
slack-bolt[async]==1.28.0
    # via mindforge (pyproject.toml)
slack-sdk==3.41.0
    # via slack-bolt
sniffio==1.3.1
    # via openai
sqlalchemy[asyncio]==2.0.49
    # via
    #   alembic
    #   mindforge (pyproject.toml)
starlette==1.0.0
    # via fastapi
tiktoken==0.12.0
    # via litellm
tokenizers==0.22.2
    # via litellm
tqdm==4.67.3
    # via
    #   huggingface-hub
    #   openai
typer==0.23.1
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   huggingface-hub
    #   minio
    #   openai
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via
    #   minio
    #   requests
uvicorn[standard]==0.44.0
    # via mindforge (pyproject.toml)
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
wrapt==1.17.3
    # via langfuse
yarl==1.23.0
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
